Move HTMLgen classes into own package
no other changes.
